Add the first demo - TurnstileStateMachine
Added the concept of a Readable/ReadOnlyStateMachine so that transitions can be hidden and business logic can take center stage. I think this is the best way of defining one of these engines. It's clear, safe, prevents unwanted transitions, and it's not really any more code than alternatives; you have to have this business logic somewhere.

Also bumped the JDK / JRE to 17. Record classes are a really nice way to implement the states, so let's utilize them. Considered enabling preview features so I could use pattern matching. This would make the business logic even cleaner and clearer, but I'll save that for another day.